+++ core-ng/src/main/java/core/framework/internal/db/DatabaseOperation.java
@@ -13,6 +13,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.sql.Types;
+import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
@@ -210,7 +211,18 @@ public class DatabaseOperation {
             // TIMESTAMP has a range of '1970-01-01 00:00:01' UTC to '2038-01-19 03:14:07' UTC.
             // for timestamp column type, there is year 2038 issue
             // for datetime column type, jdbc will save UTC values
-            statement.setTimestamp(index, Timestamp.from(((ZonedDateTime) param).toInstant()));
+
+            // with insert ignore, out of range timestamp param will be converted to "0000-00-00 00:00:00" into db, and will trigger "Zero date value prohibited" error on read
+            // refer to https://dev.mysql.com/doc/refman/8.0/en/insert.html
+            // Data conversions that would trigger errors abort the statement if IGNORE is not specified. With IGNORE, invalid values are adjusted to the closest values and inserted; warnings are produced but the statement does not abort.
+
+            // here only to check > 0, make trade off between validating TIMESTAMP column type and keeping compatible with DATETIME column type
+            // most likely the values we deal with from external systems are lesser (e.g. nodejs default year is 1900, it converts 0 into 1900/01/01 00:00:00)
+            // if it passes timestamp after 2038-01-19 03:14:07 (Instant.ofEpochSecond(Integer.MAX_VALUE)), it will still trigger this issue on MySQL
+            // so on application level, if you can not ensure the range of input value, write its own utils to check
+            Instant instant = ((ZonedDateTime) param).toInstant();
+            if (instant.getEpochSecond() <= 0) throw new Error("timestamp must be after 1970-01-01 00:00:00, value=" + param);
+            statement.setTimestamp(index, Timestamp.from(instant));
         } else if (param instanceof Boolean) {
             statement.setBoolean(index, (Boolean) param);
         } else if (param instanceof Long) {